+++ plugin/sdl2/impl/application/sdl_application.h
@@ -33,7 +33,7 @@ public:
 private:
     void initialize();
 
-    void pollEvents(uint64_t timestamp);
+    void pollEvents(uint32_t timestamp);
     uint32_t toSDLWindowFlag(const sp<ApplicationContext>& applicationContext, ApplicationManifest::WindowFlags appWindowFlag);
 
 private:
