+++ src/renderer/base/buffer.cpp
@@ -24,7 +24,7 @@ public:
             uploader.write(j.buf(), static_cast<uint32_t>(j.length()), static_cast<uint32_t>(i));
     }
 
-    bool update(uint64_t /*timestamp*/) override {
+    bool update(uint32_t /*tick*/) override {
         return false;
     }
