+++ src/renderer/base/graphics_buffer_allocator.cpp
@@ -15,7 +15,7 @@ public:
         : Uploader(size) {
     }
 
-    bool update(uint64_t /*timestamp*/) override {
+    bool update(uint32_t /*tick*/) override {
         return false;
     }
