+++ src/renderer/base/texture.h
@@ -162,7 +162,7 @@ public:
 
     const sp<Uploader>& uploader() const;
 
-    bool update(uint64_t timestamp) const;
+    bool update(uint32_t timestamp) const;
 
 //  [[plugin::resource-loader]]
     class BUILDER final : public Builder<Texture> {
