+++ src/renderer/impl/importer/atlas_importer_max_rects.cpp
@@ -23,7 +23,7 @@ public:
 
     }
 
-    bool update(uint64_t /*timestamp*/) override {
+    bool update(uint32_t /*tick*/) override {
         return false;
     }
